feat: Add user popover on explore page
